tests/wpt/web-platform-tests/2dcontext/shadows/2d.shadow.filter.value.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /2dcontext/tools/gentest.py. -->
<title>Canvas test: 2d.shadow.filter.value</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.filter.value</h1>
<p class="desc">test if ctx.filter works correctly</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("test if ctx.filter works correctly");
_addTest(function(canvas, ctx) {

_assert(ctx.filter == 'none', "ctx.filter == 'none'");
ctx.filter = 'blur(5px)';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
ctx.save();
ctx.filter = 'none';
_assert(ctx.filter == 'none', "ctx.filter == 'none'");
ctx.restore();
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

ctx.filter = 'blur(10)';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
ctx.filter = 'blur 10px';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

ctx.filter = 'inherit';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
ctx.filter = 'initial';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
ctx.filter = 'unset';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

ctx.filter = '';
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
ctx.filter = null;
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
ctx.filter = undefined;
_assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

ctx.filter = 'blur(  5px)';
assert_equals(ctx.filter, 'blur(  5px)');


});
</script>
